requirements.txt
Django~=3.2.16
pytz==2022.7
sqlparse==0.4.3

djangorestframework~=3.14.0
django-ckeditor~=6.3.2
dadata~=21.10.1